springboot_appl

+++ src/main/java/com/Springboot/EAMS/service/EmployeeService.java
@@ -1,16 +1,19 @@
 package com.Springboot.EAMS.service;
 
-import com.Springboot.EAMS.dto.EmployeeDTO;
-import com.Springboot.EAMS.entity.Employee;
+import com.Springboot.EAMS.exception.GlobalEamsException;
+import com.Springboot.EAMS.model.dto.EmployeeDTO;
+import com.Springboot.EAMS.model.entity.Employee;
+import com.Springboot.EAMS.model.entity.EmployeeDetails;
 import com.Springboot.EAMS.repo.EmployeeRepo;
-import com.Springboot.EAMS.exception.EamsException;
+import com.Springboot.EAMS.repo.testint;
 import org.modelmapper.ModelMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
+import javax.annotation.PostConstruct;
 import java.util.Optional;
 
 
@@ -36,6 +39,9 @@ public class EmployeeService {
     @Autowired
     private ModelMapper modelMapper;
 
+    @Autowired
+    EmployeeDetails empdetails;
+
 
     /*private PostDto convertToDto(Post post) {
        PostDto postDto = modelMapper.map(post, PostDto.class);
@@ -46,41 +52,54 @@ public class EmployeeService {
 
     //create
     public Employee save(EmployeeDTO employee_dto){
-
+        empdetails=employee_dto.getEmployeeDetails();
+        Employee emp = modelMapper.map(employee_dto,Employee.class);
+        emp.setEmployeeDetails(empdetails);
+        empdetails.setEmployee(emp);
         //logger.info("The Request Object is : %s",employee_dto.toString());
-        return  repo.save(modelMapper.map(employee_dto,Employee.class));
+        repo.save(emp);
+         return emp;
     }
+
     //retrieve
     public Employee get(long id){
         Optional<Employee> employee = repo.findById(id);
         if (!employee.isPresent())
-            throw new EamsException("NOT FOUND id-" + id);
+            throw new GlobalEamsException("NOT FOUND employee id-" + id);
         // Initiating Logging for Get Request
         return employee.get();
     }
+
     //delete
     public void delete(long id){
+        Optional<Employee> employee = repo.findById(id);
+        if (!employee.isPresent())
+            throw new GlobalEamsException("NOT FOUND employee id-" + id +" Delete not possible");
         repo.deleteById(id);
     }
+
     //update
     public Employee update(EmployeeDTO employee_dto, long id){
         Optional<Employee> employee = repo.findById(id);
-
         if (!employee.isPresent())
-            throw new EamsException("NOT FOUND id-" + id);
-
-
+            throw new GlobalEamsException("NOT FOUND id-" + id);
         return repo.save(modelMapper.map(employee_dto,Employee.class));
-
     }
-    public void updateallsalary(long increment){
-        if (increment<=0)
-                return;
 
+    /*ublic void updateallsalary(long increment){
+        if (increment<=0)
+                return ;
         repo.updatesalary(increment);
 
+    }*/
 
+    @Autowired
+    testint init;
+    @PostConstruct
+    public void init(){
+        init.findname(1);
     }
+
 
 
 }
