+++ src/main/java/com/Springboot/EAMS/service/OrganisationService.java
@@ -1,10 +1,10 @@
 package com.Springboot.EAMS.service;
 
 
-import com.Springboot.EAMS.dto.OrganisationDTO;
-import com.Springboot.EAMS.entity.Organisation;
+import com.Springboot.EAMS.exception.GlobalEamsException;
+import com.Springboot.EAMS.model.dto.OrganisationDTO;
+import com.Springboot.EAMS.model.entity.Organisation;
 import com.Springboot.EAMS.repo.OrganisationRepo;
-import com.Springboot.EAMS.exception.EamsException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -33,29 +33,30 @@ public class OrganisationService {
 
     //create
     public Organisation save(OrganisationDTO dto){
-
         return  repo.save(modelMapper.map(dto,Organisation.class));
     }
+
     //retrieve
     public Organisation get(long id){
         Optional<Organisation> organisation= repo.findById(id);
         if (!organisation.isPresent())
-            throw new EamsException("NOT FOUND id-" + id);
+            throw new GlobalEamsException("NOT FOUND organisation id-" + id);
 
         return organisation.get();
     }
     //delete
     public void delete(long id){
+        Optional<Organisation> organisation = repo.findById(id);
+        if (!organisation.isPresent())
+            throw new GlobalEamsException("NOT FOUND organisation id-" + id);
         repo.deleteById(id);
     }
 
     //update
     public Organisation update(OrganisationDTO dto, long id){
         Optional<Organisation> organisation = repo.findById(id);
-
         if (!organisation.isPresent())
-            throw new EamsException("NOT FOUND id-" + id);
-
+            throw new GlobalEamsException("NOT FOUND id organisation-" + id);
         return repo.save(modelMapper.map(dto,Organisation.class));
 
     }
